+++ Source/TerritorialPissings/tpcChien.cpp
@@ -2,13 +2,14 @@
 
 
 #include "tpcChien.h"
+#include "tpcPromeneuse.h"
 
 
 void AtpcChien::BeginPlay()
 {
 	Super::BeginPlay();
 
-	GEngine->AddOnScreenDebugMessage(-1, 5.f, FColor::Red, TEXT("Screen Message"));
+	aPromeneur = UGameplayStatics::GetActorOfClass(GetWorld(), AtpcPromeneuse::StaticClass());
 }
 
 void AtpcChien::SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent)
@@ -17,6 +18,8 @@ void AtpcChien::SetupPlayerInputComponent(class UInputComponent* PlayerInputComp
 	check(PlayerInputComponent);
 	PlayerInputComponent->BindAction("Jump", IE_Pressed, this, &ACharacter::Jump);
 	PlayerInputComponent->BindAction("Jump", IE_Released, this, &ACharacter::StopJumping);
+
+	PlayerInputComponent->BindAction("Sprint", IE_Pressed, this, &AtpcChien::Sprint);
 
 	PlayerInputComponent->BindAxis("MoveForward", this, &AtpcChien::MoveForward);
 	PlayerInputComponent->BindAxis("MoveRight", this, &AtpcChien::MoveRight);
@@ -48,7 +51,30 @@ void AtpcChien::MoveForward(float Value)
 
 		// get forward vector
 		const FVector Direction = FRotationMatrix(YawRotation).GetUnitAxis(EAxis::X);
-		AddMovementInput(Direction, Value);
+
+		if (!IsValid(aPromeneur))
+		{
+			GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Red, TEXT("Promeneur invalide"));
+			return;
+		}
+
+		const FVector vLocation = this->GetActorLocation();
+		const FVector vCible = aPromeneur->GetActorLocation();
+		const FVector vResult = (vLocation + (Direction*Value));
+
+		float fDistance = FVector::Distance(vLocation, vCible);
+		float fResult = FVector::Distance(vResult, vCible);
+
+		if (fResult <= 1000 || fResult < fDistance || bSprint)
+		{
+			GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Red, TEXT("Mouvement"));
+			AddMovementInput(Direction, Value);
+		}
+		else
+		{
+
+			GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Red, TEXT("MouvementImpossible"));
+		}
 	}
 }
 
@@ -63,6 +89,43 @@ void AtpcChien::MoveRight(float Value)
 		// get right vector 
 		const FVector Direction = FRotationMatrix(YawRotation).GetUnitAxis(EAxis::Y);
 		// add movement in that direction
-		AddMovementInput(Direction, Value);
+
+		if (!IsValid(aPromeneur))
+		{
+			GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Red, TEXT("Promeneur invalide"));
+			return;
+		}
+
+		const FVector vLocation = this->GetActorLocation();
+		const FVector vCible = aPromeneur->GetActorLocation();
+		const FVector vResult = (vLocation + (Direction * Value));
+
+		float fDistance = FVector::Distance(vLocation, vCible);
+		float fResult = FVector::Distance(vResult, vCible);
+
+		//Gestion de l'éloignement avec la laisse.
+		if (fResult <= 1000 || fResult < fDistance || bSprint)
+		{
+			GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Red, TEXT("Mouvement"));
+			AddMovementInput(Direction, Value);
+		}
 	}
+}
+
+
+void AtpcChien::Sprint()
+{
+	GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Red, TEXT("Sprint"));
+	bSprint = true;
+	
+	FTimerHandle UniqueHandle;
+	FTimerDelegate SprintDelegate = FTimerDelegate::CreateUObject(this, &AtpcChien::StopSprint);
+	GetWorldTimerManager().SetTimer(UniqueHandle, SprintDelegate, 1.f, false);
+}
+
+
+void AtpcChien::StopSprint()
+{
+	GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Red, TEXT("StopSprint"));
+	bSprint = false;
 }