+++ Source/TerritorialPissings/tpcChien.h
@@ -4,6 +4,8 @@
 
 #include "CoreMinimal.h"
 #include "TerritorialPissingsCharacter.h"
+#include "Kismet/GameplayStatics.h"
+#include "Engine/EngineTypes.h"
 #include "tpcChien.generated.h"
 
 /**
@@ -22,7 +24,14 @@ protected :
 	/** Called for side to side input */
 	void MoveRight(float Value);
 
+	void Sprint();
+	void StopSprint();
+
 protected:
+
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Sprint, meta = (AllowPrivateAccess = "true"))
+	bool bSprint = false;
+
 	// APawn interface
 	virtual void SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent) override;
 	// End of APawn interface
@@ -30,6 +39,6 @@ protected:
 	void BeginPlay();
 
 private:
-	//ACharacter cPromeneur;
+	AActor* aPromeneur;
 	
 };